fix server api

+++ src/app/api/characterDelete/route.ts
@@ -3,6 +3,9 @@ import { sql } from '@vercel/postgres'
 export async function DELETE(req: Request) {
   const url = new URL(req.url)
   const characterName = url.searchParams.get('characterName')
+  if (!characterName) {
+    return new Response(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 404 })
+  }
 
   try {
     const res = await sql`DELETE FROM characters WHERE character_name=${characterName}`
